fix stream types

+++ src/android/VolumeChangeCallback.java
@@ -4,5 +4,5 @@ package com.datamart.wfpk;
  * Created by r.stepanyak on 9/7/15.
  */
 public interface VolumeChangeCallback {
-    void volumeChanged(double newVolume);
+    void volumeChanged(float newVolume);
 }
